Set $LOAD_PATH for the application
This uses Kernel#caller (which is fragile) to determine where Rulers was
loaded from and determine what load paths to add (like app/controllers/)
automatically.

It's unfortunate that the user can't "just require" this file and needs
to call Rulers.app to initialize the application and run this, but I
suspect a future version will need something like this.